+++ src/Components/racingCarGame.js
@@ -1,22 +1,10 @@
+import { countMoves } from "../Manager/gameManager.js";
+
 export default function RacingCarGame() {
   this.play = function (cars) {
-    let winner = [];
-
-    winner.push(cars[0]);
-
-    for (let i = 1; i < cars.length; i++) {
-      if (cars[i].move.length > winner[0].move.length) {
-        winner = [];
-        winner.push(cars[i]);
-      } else if (cars[i].move.length == winner[0].move.length) {
-        winner.push(cars[i]);
-      } else {
-        continue;
-      }
-    }
-
+    const winner = countMoves(cars);
     let winnerNames = [];
-    winner.forEach((x) => winnerNames.push(x.name));
+    winner.forEach((car) => winnerNames.push(car.name));
 
     return winnerNames.join(", ");
   };
